requirements.txt
requests==2.27.1
json5==0.9.6
pyspark==3.4.1
Faker
kafka-python==2.0.2
